test updates

+++ src/kinetic_ui/handler/vh_cv_scrl.c
@@ -11,7 +11,7 @@ typedef struct _vh_cv_scrl_t
     ku_view_t* hori_v;
     int        state; // 0 scroll 1 open 2 close
     int        steps;
-    uint32_t   item_cnt;
+    size_t     item_cnt;
     void*      userdata;
 } vh_cv_scrl_t;
 
@@ -23,7 +23,7 @@ void vh_cv_scrl_attach(
 void vh_cv_scrl_update(ku_view_t* view);
 void vh_cv_scrl_show(ku_view_t* view);
 void vh_cv_scrl_hide(ku_view_t* view);
-void vh_cv_scrl_set_item_count(ku_view_t* view, uint32_t count);
+void vh_cv_scrl_set_item_count(ku_view_t* view, size_t count);
 void vh_cv_scrl_scroll_v(ku_view_t* view, int y);
 void vh_cv_scrl_scroll_h(ku_view_t* view, int x);
 
@@ -63,7 +63,7 @@ void vh_cv_scrl_attach(
     view->evt_han_data = vh;
 }
 
-void vh_cv_scrl_set_item_count(ku_view_t* view, uint32_t count)
+void vh_cv_scrl_set_item_count(ku_view_t* view, size_t count)
 {
     vh_cv_scrl_t* vh = view->evt_han_data;
